Maintainability: libs.net.build improved by moving GUI callable methods to separate modules
